*move Optimizer member implementations to separate source files.
