+++ prj/vs2022/CvtCore.vcxproj
@@ -27,13 +27,22 @@
     <ClCompile Include="..\..\src\Cvt\Deoptimizer\Deoptimizer.cpp" />
     <ClCompile Include="..\..\src\Cvt\InferenceEngine\InferenceEngine.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeActivation.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeBiasAndScale.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeConvolution.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeConvolutionAndPower.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeConvolutionAndScale.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeCurrentAndBias.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeInnerProductAndPower.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeInnerProductAndScale.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergePowerAndScaleAndPower.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeQuantized.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeQuantizedSqueezeExcitation.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeShuffle.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeSoftmax.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeSqueezeExcitation.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\Optimizer.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\ReduceTensorIteratorIO.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\TransposeInnerProduct.cpp" />
     <ClInclude Include="..\..\src\Cvt\Common\Params.h" />
     <ClInclude Include="..\..\src\Cvt\Common\SynetUtils.h" />
     <ClInclude Include="..\..\src\Cvt\Deoptimizer\Deoptimizer.h" />
