+++ src/Cvt/Optimizer/MergeBiasAndScale.cpp
@@ -1,0 +1,61 @@
+/*
+* Synet Framework (http://github.com/ermig1979/Synet).
+*
+* Copyright (c) 2018-2025 Yermalayeu Ihar.
+*
+* Permission is hereby granted, free of charge, to any person obtaining a copy
+* of this software and associated documentation files (the "Software"), to deal
+* in the Software without restriction, including without limitation the rights
+* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+* copies of the Software, and to permit persons to whom the Software is
+* furnished to do so, subject to the following conditions:
+*
+* The above copyright notice and this permission notice shall be included in
+* all copies or substantial portions of the Software.
+*
+* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+* SOFTWARE.
+*/
+
+#include "Cvt/Optimizer/Common.h"
+#include "Cvt/Optimizer/Optimizer.h"
+
+namespace Synet
+{
+    bool MergeBiasAndScale(const LayerParams& src, size_t& index, Bytes& bin, Bytes& buf, LayerParams& dst, Changes& changes)
+    {
+        if (index == 0)
+            return false;
+        const LayerParam & bias = src[index - 1];
+        const LayerParam & scale = src[index];
+        if (bias.type() != LayerTypeBias)
+            return false;
+        if (scale.type() != LayerTypeScale || scale.scale().biasTerm() || scale.src()[0] != bias.dst()[0])
+            return false;
+        if (InsideLink(src, index - 1, 2))
+            return false;
+        dst.back().type() = LayerTypeScale;
+        dst.back().name() = scale.name();
+        dst.back().dst() = scale.dst();
+        dst.back().scale().biasTerm() = true;
+        dst.back().weight().push_back(scale.weight()[0]);
+        if (buf.empty())
+            buf = bin;
+        const float* pSrcScale = GetWeight<float>(bin, dst.back().weight()[1]);
+        const float* pSrcBias = GetWeight<float>(bin, dst.back().weight()[0]);
+        float* pDstScale = GetWeight<float>(buf, dst.back().weight()[0]);
+        float* pDstBias = GetWeight<float>(buf, dst.back().weight()[1]);
+        size_t size = TensorSize(dst.back().weight()[0].dim());
+        for (size_t i = 0; i < size; ++i)
+        {
+            pDstScale[i] = pSrcScale[i];
+            pDstBias[i] = pSrcScale[i] * pSrcBias[i];
+        }
+        return true;
+    }
+}
