+++ src/Cvt/Optimizer/Optimizer.h
@@ -46,529 +46,6 @@ namespace Synet
 
         bool OptimizeLayers(Synet::NetworkParam& network, Bytes& bin, int stage);
 
-        bool ReduceTensorIteratorIO(const LayerParams& src, size_t& index, const Bytes& bin, Bytes& buf, LayerParams& dst)
-        {
-            const LayerParam & stt = src[index];
-            if (stt.type() != LayerTypeTensorIterator || stt.src().size() < 3 || stt.tensorIterator().back().size() < 1)
-                return false;
-            size_t srcDupls = 0;
-            for (size_t i = 2; i < stt.src().size(); ++i)
-            {
-                if (stt.src()[1] == stt.src()[i])
-                    srcDupls++;
-            }
-            size_t backDupls = 0;
-            for (size_t i = 1; i < stt.tensorIterator().back().size(); ++i)
-            {
-                if (stt.tensorIterator().back()[0].src() == stt.tensorIterator().back()[i].src())
-                    backDupls++;
-            }
-            if (srcDupls == 0 || srcDupls != backDupls || srcDupls < stt.src().size() - 2)
-                return false;
-            dst.push_back(stt);
-            LayerParam& dtt = dst.back();
-            dtt.src().resize(2);
-            String rem, iter;
-            for (size_t i = 0; i < dtt.tensorIterator().input().size() && iter.empty(); ++i)
-                if (dtt.tensorIterator().input()[i].axis() != -1)
-                    iter = dtt.tensorIterator().input()[i].dst();
-            for (size_t i = index + 1; i < src.size() && rem.empty(); ++i)
-            {
-                if (src[i].parent() != stt.name())
-                    break;
-                if (src[i].type() == LayerTypeInput && src[i].name() != iter)
-                    rem = src[i].name();
-            }
-            StringSet del;
-            std::vector<ConnectionParam> back, input;
-            for (size_t i = 0; i < dtt.tensorIterator().input().size(); ++i)
-            {
-                ConnectionParam & p = dtt.tensorIterator().input()[i];
-                if (p.dst() == rem || p.dst() == iter)
-                {
-                    p.port() = Synet::Min<int>(1, p.port());
-                    input.push_back(p);
-                }
-                else
-                    del.insert(p.dst());
-            }
-            dtt.tensorIterator().input().swap(input);
-            for (size_t i = 0; i < stt.tensorIterator().back().size(); ++i)
-            {
-                if (del.find(dtt.tensorIterator().back()[i].dst()) == del.end())
-                    back.push_back(dtt.tensorIterator().back()[i]);
-            }
-            dtt.tensorIterator().back().swap(back);
-            for (size_t i = index + 1; i < src.size(); ++i)
-            {
-                if (src[i].parent() != stt.name())
-                    break;
-                if (src[i].type() != LayerTypeInput || del.find(src[i].name()) == del.end())
-                    dst.push_back(src[i]);
-                for (size_t j = 0; j < dst.back().src().size(); ++j)
-                {
-                    if (del.find(dst.back().src()[j]) != del.end())
-                        dst.back().src()[j] = rem;
-                }
-                index++;
-            }
-            return true;
-        }
-
-        bool TransposeInnerProduct(const LayerParams& src, size_t& index, const Bytes& bin, Bytes& buf, LayerParams& dst)
-        {
-            const LayerParam& ip = src[index];
-            if (ip.type() != LayerTypeInnerProduct || !ip.innerProduct().transposeB() || ip.weight().empty())
-                return false;
-            if (WeightUserCount(src, ip.weight()[0]) > 1)
-                return false;
-            const Shape & dim = ip.weight()[0].dim();
-            if (buf.empty())
-                buf = bin;
-            dst.push_back(ip);
-            dst.back().innerProduct().transposeB() = false;
-            int axis = ip.innerProduct().axis(), dim0 = (int)dim[axis - 1], dim1 = (int)dim[axis];
-            dst.back().weight()[0].dim() = Shp(dim1, dim0);
-            const float* pSrc = GetWeight<float>(bin, ip.weight()[0]);
-            float* pDst = GetWeight<float>(buf, ip.weight()[0]);
-            for (int i = 0; i < dim0; ++i)
-                for (int j = 0; j < dim1; ++j)
-                    pDst[j * dim0 + i] = pSrc[i * dim1 + j];
-            return true;
-        }
-
-        bool MergeCurrentAndBias(const LayerParams& src, size_t& index, Bytes& bin, LayerParams& dst, Changes& changes)
-        {
-            if (index == 0)
-                return false;
-            const LayerParam & current = src[index - 1];
-            const LayerParam & bias = src[index];
-            const WeightParam* weight = GetEltwiseWeight(index, src);
-            if (bias.type() == LayerTypeBias)
-                weight = &bias.weight()[0];
-            if(weight == NULL)
-                return false;
-            if (!(bias.src()[0] == current.name() || (bias.src().size() == 2 && bias.src()[1] == current.name())))
-                return false;
-            if (InsideLink(src, index - 1, 2))
-                return false;
-            switch (current.type())
-            {
-            case LayerTypeConvolution:
-                if (current.convolution().biasTerm() || current.convolution().outputNum() != weight->dim()[0])
-                    return false;
-                dst.back().convolution().biasTerm() = true;
-                break;
-            case LayerTypeInnerProduct:
-                if (current.innerProduct().biasTerm() || current.src().size() != 1 || current.innerProduct().outputNum() != weight->dim()[0])
-                    return false;
-                dst.back().innerProduct().biasTerm() = true;
-                break;
-            case LayerTypePower:
-                if (current.power().power() != 1.0f || current.power().shift() != 0.0f || bias.type() != LayerTypeBias)
-                    return false;
-                dst.back().type() = LayerTypeScale;
-                dst.back().scale().biasTerm() = true;
-                dst.back().weight().push_back(*weight);
-                dst.back().weight()[0].offset() = bin.size();
-                for (size_t i = 0; i < dst.back().weight()[0].dim()[0]; ++i)
-                    PushBack<float>(bin, current.power().scale());
-                dst.back().power().scale() = 1.0f;
-                break;
-            case LayerTypeScale:
-                if (current.scale().biasTerm())
-                    return false;
-                dst.back().scale().biasTerm() = true;
-                break;
-                default:
-                    return false;
-            }
-            dst.back().name() = bias.name();
-            dst.back().dst() = bias.dst();
-            dst.back().weight().push_back(*weight);
-            return true;
-        }
-
-        bool MergePowerAndScaleAndPower(const LayerParams& src, size_t& index, Bytes& bin, Bytes& buf, LayerParams& dst, Changes& changes)
-        {
-            bool pre = false, scale = false, post = false;
-            if (src.size() > index + 0 && src[index + 0].type() == LayerTypePower && src[index + 0].power().power() == 1.0f)
-                pre = true;
-            if (src.size() > index + 1 && src[index + 1].type() == LayerTypeScale && (pre ? src[index + 1].src()[0] == src[index + 0].name() : true) && src[index + 1].scale().biasTerm())
-                scale = true;
-            if (src.size() > index + 2 && src[index + 2].type() == LayerTypePower && src[index + 2].power().power() == 1.0f && src[index + 2].src()[0] == src[index + 1].name())
-                post = true;
-            if (!(scale && (pre || post)))
-                return false;
-            if (InsideLink(src, index + (pre ? 0 : 1), 1 + (pre ? 1 : 0) + (post ? 1 : 0), 0, LayerTypes({ LayerTypePriorBox, LayerTypePriorBoxClustered, LayerTypeMeta })))
-                return false;
-            LayerParam layer;
-            layer.type() = LayerTypeScale;
-            layer.name() = src[index + 1].name();
-            layer.src().push_back(pre ? src[index + 0].src()[0] : src[index + 1].src()[0]);
-            layer.dst().push_back(post ? src[index + 2].dst()[0] : src[index + 1].dst()[0]);
-            layer.scale() = src[index + 1].scale();
-            layer.weight() = src[index + 1].weight();
-            float preScale = pre ? src[index + 0].power().scale() : 1.0f;
-            float preBias = pre ? src[index + 0].power().shift() : 0.0f;
-            float postScale = post ? src[index + 2].power().scale() : 1.0f;
-            float postBias = post ? src[index + 2].power().shift() : 0.0f;
-            if (buf.empty())
-                buf = bin;
-            float* pScale = GetWeight<float>(buf, layer.weight()[0]);
-            float* pShift = GetWeight<float>(buf, layer.weight()[1]);
-            size_t size = TensorSize(layer.weight()[0].dim());
-            for (size_t i = 0; i < size; ++i)
-            {
-                pShift[i] = (preBias * pScale[i] + pShift[i]) * postScale + postBias;
-                pScale[i] = preScale * pScale[i] * postScale;
-            }
-            if (pre)
-                changes.push_back(Change(src[index + 0].dst()[0], layer.dst()[0]));
-            else
-                dst.push_back(src[index + 0]);
-            index += 1 + (post ? 1 : 0);
-            dst.push_back(layer);
-            return true;
-        }
-
-        bool MergeBiasAndScale(const LayerParams& src, size_t& index, Bytes& bin, Bytes& buf, LayerParams& dst, Changes& changes)
-        {
-            if (index == 0)
-                return false;
-            const LayerParam & bias = src[index - 1];
-            const LayerParam & scale = src[index];
-            if (bias.type() != LayerTypeBias)
-                return false;
-            if (scale.type() != LayerTypeScale || scale.scale().biasTerm() || scale.src()[0] != bias.dst()[0])
-                return false;
-            if (InsideLink(src, index - 1, 2))
-                return false;
-            dst.back().type() = LayerTypeScale;
-            dst.back().name() = scale.name();
-            dst.back().dst() = scale.dst();
-            dst.back().scale().biasTerm() = true;
-            dst.back().weight().push_back(scale.weight()[0]);
-            if (buf.empty())
-                buf = bin;
-            const float* pSrcScale = GetWeight<float>(bin, dst.back().weight()[1]);
-            const float* pSrcBias = GetWeight<float>(bin, dst.back().weight()[0]);
-            float* pDstScale = GetWeight<float>(buf, dst.back().weight()[0]);
-            float* pDstBias = GetWeight<float>(buf, dst.back().weight()[1]);
-            size_t size = TensorSize(dst.back().weight()[0].dim());
-            for (size_t i = 0; i < size; ++i)
-            {
-                pDstScale[i] = pSrcScale[i];
-                pDstBias[i] = pSrcScale[i] * pSrcBias[i];
-            }
-            return true;
-        }
-
-        bool MergeConvolutionAndPower(const LayerParams& src, size_t& index, const Bytes& bin, Bytes& buf, LayerParams& dst, Changes& changes)
-        {
-            if (index == 0)
-                return false;
-            const LayerParam& conv = src[index - 1];
-            const LayerParam& power = src[index];
-            if (conv.type() != LayerTypeConvolution || 
-                conv.convolution().activationType() != ActivationFunctionTypeIdentity)
-                return false;
-            if (power.type() != LayerTypePower || power.src()[0] != conv.name() ||
-                power.power().power() != 1.0f || power.power().shift() != 0.0f)
-                return false;
-            if (InsideLink(src, index - 1, 2))
-                return false;
-            if (conv.weight()[0].format() != TensorFormatNhwc)
-                return false;
-            if (buf.empty())
-                buf = bin;
-            dst.back().name() = power.name();
-            dst.back().dst() = power.dst();
-            float scale = power.power().scale();
-            for (size_t w = 0; w < conv.weight().size(); ++w)
-            {
-                const float* pSrc = GetWeight<float>(bin, conv.weight()[w]);
-                float* pDst = GetWeight<float>(buf, conv.weight()[w]);
-                for (size_t i = 0, n = conv.weight()[w].size() / 4; i < n; ++i)
-                    pDst[i] = pSrc[i] * scale;
-            }
-            return true;
-        }
-
-        bool MergeInnerProductAndPower(const LayerParams& src, size_t& index, const Bytes& bin, Bytes& buf, LayerParams& dst, Changes& changes)
-        {
-            if (index == 0)
-                return false;
-            const LayerParam& ip = src[index - 1];
-            const LayerParam& power = src[index];
-            if (ip.type() != LayerTypeInnerProduct || ip.src().size() != 1)
-                return false;
-            if (power.type() != LayerTypePower || power.src()[0] != ip.name() ||
-                power.power().power() != 1.0f || power.power().shift() != 0.0f)
-                return false;
-            if (InsideLink(src, index - 1, 2))
-                return false;
-            if (buf.empty())
-                buf = bin;
-            dst.back().name() = power.name();
-            dst.back().dst() = power.dst();
-            float scale = power.power().scale();
-            for (size_t w = 0; w < ip.weight().size(); ++w)
-            {
-                const float* pSrc = GetWeight<float>(bin, ip.weight()[w]);
-                float* pDst = GetWeight<float>(buf, ip.weight()[w]);
-                for (size_t i = 0, n = ip.weight()[w].size() / 4; i < n; ++i)
-                    pDst[i] = pSrc[i] * scale;
-            }
-            return true;
-        }
-
-        bool MergeSoftmax(const LayerParams & src, size_t & index, LayerParams & dst, Changes & changes)
-        {
-            if (index == 0 || src.size() < index + 5)
-                return false;
-            if (src[index + 0].type() != LayerTypeReduction || src[index + 0].reduction().type() != ReductionTypeMax ||
-                src[index + 0].reduction().axis().size() != 1)
-                return false;
-            if (src[index + 1].type() != LayerTypeBinaryOperation || src[index + 1].binaryOperation().type() != BinaryOperationTypeSub ||
-                src[index + 1].src()[0] != src[index + 0].src()[0] || src[index + 1].src()[1] != src[index + 0].name())
-                return false;
-            if (src[index + 2].type() != LayerTypeUnaryOperation || src[index + 2].unaryOperation().type() != UnaryOperationTypeExp ||
-                src[index + 2].src()[0] != src[index + 1].name())
-                return false;
-            if (src[index + 3].type() != LayerTypeReduction || src[index + 3].reduction().type() != ReductionTypeSum ||
-                src[index + 3].reduction().axis() != src[index + 0].reduction().axis() || src[index + 3].src()[0] != src[index + 2].name())
-                return false;
-            if (src[index + 4].type() != LayerTypeBinaryOperation || src[index + 4].binaryOperation().type() != BinaryOperationTypeDiv ||
-                src[index + 4].src()[0] != src[index + 2].name() || src[index + 4].src()[1] != src[index + 3].name())
-                return false;
-            for (size_t i = index + 5; i < src.size(); ++i)
-            {
-                for (size_t j = 0; j < src[i].src().size(); ++j)
-                {
-                    for (ptrdiff_t k = 0; k < 4; ++k)
-                    {
-                        if (src[i].src()[j] == src[index + k].name())
-                            return false;
-                    }
-                }
-            }
-            LayerParam layer;
-            layer.type() = LayerTypeSoftmax;
-            layer.name() = src[index + 4].name();
-            layer.src().push_back(src[index + 0].src()[0]);
-            layer.dst().push_back(layer.name());
-            layer.softmax().axis() = src[index + 0].reduction().axis()[0];
-            dst.push_back(layer);
-            index += 4;
-            return true;
-        }
-
-        bool MergeShuffle0(const LayerParams & src, size_t & index, LayerParams & dst, Changes & changes)
-        {
-            if (src.size() < index + 5)
-                return false;
-            if (src[index + 0].type() != LayerTypeConcat || src[index + 0].src().size() != 2)
-                return false;
-            if (src[index + 1].type() != LayerTypeReshape || src[index + 1].reshape().shape().size() != 3)
-                return false;
-            if (src[index + 2].type() != LayerTypePermute)
-                return false;
-            if (src[index + 3].type() != LayerTypeUnpack || src[index + 3].dst().size() != 2)
-                return false;
-            if (src[index + 4].type() != LayerTypeReshape || 
-                src[index + 4].reshape().shape().size() + src[index + 4].reshape().axis() != 4)
-                return false;
-            if (src[index + 5].type() != LayerTypeReshape || 
-                src[index + 5].reshape().shape().size() + src[index + 5].reshape().axis() != 4)
-                return false;
-            if (InsideLink(src, index, 4, 1))
-                return false;
-            LayerParam layer;
-            layer.type() = LayerTypeShuffle;
-            layer.name() = src[index + 0].name();
-            layer.src() = src[index + 0].src();
-            layer.shuffle().type() = 0;
-            layer.dst().push_back(src[index + 4].dst()[0]);
-            layer.dst().push_back(src[index + 5].dst()[0]);
-            index += 5;
-            dst.push_back(layer);
-            return true;
-        }
-
-        bool MergeShuffle1(const LayerParams & src, size_t & index, LayerParams & dst, Changes & changes)
-        {
-            if (src.size() < index + 4)
-                return false;
-            if (src[index + 0].type() != LayerTypeConcat || src[index + 0].src().size() != 2)
-                return false;
-            if (src[index + 1].type() != LayerTypeReshape || src[index + 1].reshape().axis() + src[index + 1].reshape().shape().size() != 5)
-                return false;
-            if (src[index + 2].type() != LayerTypePermute)
-                return false;
-            if (src[index + 3].type() != LayerTypeReshape || src[index + 3].reshape().axis() + src[index + 3].reshape().shape().size() != 4)
-                return false;
-            if (src[index + 4].type() != LayerTypeUnpack || src[index + 4].dst().size() != 2)
-                return false;
-            if (InsideLink(src, index, 4, 0))
-                return false;
-            LayerParam layer;
-            layer.type() = LayerTypeShuffle;
-            layer.name() = src[index + 0].name();
-            layer.src() = src[index + 0].src();
-            layer.shuffle().type() = 1;
-            layer.dst().push_back(src[index + 4].dst()[0]);
-            layer.dst().push_back(src[index + 4].dst()[1]);
-            index += 4;
-            dst.push_back(layer);
-            return true;
-        }
-
-        bool MergeShuffle2(const LayerParams & src, size_t & index, LayerParams & dst, Changes & changes)
-        {
-            if (src.size() < index + 18)
-                return false;
-            if (src[index + 0].type() != LayerTypeConcat || src[index + 0].src().size() != 2)
-                return false;
-            if (src[index + 1].type() != LayerTypeReshape)// || src[index + 1].reshape().axis() + src[index + 1].reshape().shape().size() != 5)
-                return false;
-            if (src[index + 2].type() != LayerTypePermute || src[index + 2].permute().order().size() != 5)
-                return false;
-            if (src[index + 3].type() != LayerTypeReshape)// || src[index + 3].reshape().axis() + src[index + 3].reshape().shape().size() != 4)
-                return false;
-            for (size_t i = 4; i < 14; ++i)
-            {
-                if (src[index + i].type() != LayerTypeMeta)
-                    return false;
-            }
-            if (src[index + 14].type() != LayerTypeStridedSlice || src[index + 14].src().size() != 4)
-                return false;
-            for (size_t i = 15; i < 16; ++i)
-            {
-                if (src[index + i].type() != LayerTypeMeta)
-                    return false;
-            }
-            if (src[index + 17].type() != LayerTypeStridedSlice || src[index + 17].src().size() != 4)
-                return false;
-            LayerParam layer;
-            layer.type() = LayerTypeShuffle;
-            layer.name() = src[index + 0].name();
-            layer.src() = src[index + 0].src();
-            layer.shuffle().type() = 1;
-            layer.dst().push_back(src[index + 14].dst()[0]);
-            layer.dst().push_back(src[index + 17].dst()[0]);
-            index += 17;
-            dst.push_back(layer);
-            return true;
-        }
-
-        bool MergeShuffle3(const LayerParams& src, size_t& index, LayerParams& dst, Changes& changes)
-        {
-            if (src.size() < index + 43)
-                return false;
-            if (src[index + 0].type() != LayerTypeConcat || src[index + 0].src().size() != 2)
-                return false;
-            for(size_t i = 1; i < 22; ++i)
-                if (src[index + i].type() != LayerTypeMeta)
-                    return false;
-            if (src[index + 22].type() != LayerTypeReshape)
-                return false;
-            if (src[index + 23].type() != LayerTypePermute || src[index + 23].permute().order().size() != 5)
-                return false;
-            for (size_t i = 24; i < 28; ++i)
-                if (src[index + i].type() != LayerTypeMeta)
-                    return false;
-            if (src[index + 28].type() != LayerTypeReshape)
-                return false;
-            for (size_t i = 29; i < 39; ++i)
-                if (src[index + i].type() != LayerTypeMeta)
-                    return false;
-            if (src[index + 39].type() != LayerTypeStridedSlice || src[index + 39].src().size() != 4)
-                return false;
-            for (size_t i = 40; i < 42; ++i)
-                if (src[index + i].type() != LayerTypeMeta)
-                    return false;
-            if (src[index + 42].type() != LayerTypeStridedSlice || src[index + 42].src().size() != 4)
-                return false;
-            LayerParam layer;
-            layer.type() = LayerTypeShuffle;
-            layer.name() = src[index + 0].name();
-            layer.src() = src[index + 0].src();
-            layer.shuffle().type() = 1;
-            layer.dst().push_back(src[index + 39].dst()[0]);
-            layer.dst().push_back(src[index + 42].dst()[0]);
-            index += 42;
-            dst.push_back(layer);
-            return true;
-        }
-
-        bool MergeShuffle3cut(const LayerParams& src, size_t& index, bool isNhwc, LayerParams& dst, Changes& changes)
-        {
-            if (src.size() < index + 29)
-                return false;
-            if (src[index + 0].type() != LayerTypeConcat || src[index + 0].src().size() != 2)
-                return false;
-            for (size_t i = 1; i < 22; ++i)
-                if (src[index + i].type() != LayerTypeMeta)
-                    return false;
-            if (src[index + 22].type() != LayerTypeReshape)
-                return false;
-            if (src[index + 23].type() != LayerTypePermute || src[index + 23].permute().order().size() != 5)
-                return false;
-            for (size_t i = 24; i < 28; ++i)
-                if (src[index + i].type() != LayerTypeMeta)
-                    return false;
-            if (src[index + 28].type() != LayerTypeReshape)
-                return false;
-            LayerParam shuffle;
-            shuffle.type() = LayerTypeShuffle;
-            shuffle.name() = src[index + 0].name();
-            shuffle.src() = src[index + 0].src();
-            shuffle.shuffle().type() = 1;
-            shuffle.dst().push_back(src[index + 28].dst()[0] + "_dst0");
-            shuffle.dst().push_back(src[index + 28].dst()[0] + "_dst1");
-            LayerParam concat;
-            concat.type() = LayerTypeConcat;
-            concat.name() = src[index + 28].name();
-            concat.src() = shuffle.dst();
-            concat.dst().push_back(src[index + 28].dst()[0]);
-            concat.concat().axis() = isNhwc ? -1 : 1;
-            index += 28;
-            dst.push_back(shuffle);
-            dst.push_back(concat);
-            return true;
-        }
-
-        bool MergeShuffle4(const LayerParams& src, size_t& index, LayerParams& dst, Changes& changes)
-        {
-            if (src.size() < index + 5)
-                return false;
-            if (src[index + 0].type() != LayerTypeConcat || src[index + 0].src().size() != 2)
-                return false;
-            if (src[index + 1].type() != LayerTypeReshape || src[index + 1].reshape().axis() + src[index + 1].reshape().shape().size() != 5)
-                return false;
-            if (src[index + 2].type() != LayerTypePermute)
-                return false;
-            if (src[index + 3].type() != LayerTypeReshape || src[index + 3].reshape().axis() + src[index + 3].reshape().shape().size() != 4)
-                return false;
-            if (src[index + 4].type() != LayerTypeStridedSlice || src[index + 4].src()[0] != src[index + 3].dst()[0])
-                return false;
-            if (src[index + 5].type() != LayerTypeStridedSlice || src[index + 5].src()[0] != src[index + 3].dst()[0])
-                return false;
-            if (InsideLink(src, index, 4, 1))
-                return false;
-            LayerParam layer;
-            layer.type() = LayerTypeShuffle;
-            layer.name() = src[index + 0].name();
-            layer.src() = src[index + 0].src();
-            layer.shuffle().type() = 1;
-            layer.dst().push_back(src[index + 4].dst()[0]);
-            layer.dst().push_back(src[index + 5].dst()[0]);
-            index += 5;
-            dst.push_back(layer);
-            return true;
-        }
-
         bool MergePermute(const LayerParams & src, size_t & index, LayerParams & dst, Changes & changes)
         {
             if (src.size() < index + 3)
